Basic functionality works, with a caveat.
When in CODE_STATE, buttons pressed will get recognized once the board is back in MUSIC_STATE, which will cause a large amount of noise.
I think the reason is that those press events coming in aren't being "serviced" because the code is stuck in a Thread.sleep() most of the time while in CODE_STATE.
Going to be labbing with ScheduledThreadPoolExecutor to see if that fixes the issue.